Add guild invite filtering
This commit adds the invite filtering implementation to the new system.

This also fixes an issue with the implementation of the extension filtering, where there was no way to tell the bot to ignore a user when they posted a non-whitelisted file extension, since there's no relevant filter in this scenario.
Instead the extensions and invites filters now use the whitelist validation defaults to dictate when filtering should be done at all. For example, if the list validations are to ignore Helpers, then no invite filtering will occur no matter what.
The meaning of this is that the system is somewhat less configurable, because settings overrides in filters belonging to a whitelist are meaningless.

Additionally this commit contains the following fixes:
- If the user tries to show the filters in a list which doesn't exist, it says it doesn't exist instead of saying there are 0 filters.
- The filter context content is `Union[str, set]` instead of `Union[str, set[str]]`.
- An empty embed will no longer be created when `dm_embed` is empty.