Add domain filtering
The domain filtering works very similarly to the token filtering, and the domain matching itself is based on the implementation in the old system.

The deletion setting is accessed explicitly in the domain filter in order to allow DMing the user the domain the message got deleted for.
This is fine, since practical uses are more important than the theory this system was designed with. Of course, breaking the design should still be avoided whenever possible.