Add filter-type-specific settings
Pydantic is added to the bot's dependencies, allowing easily setting up settings that are specific to a filter type (e.g for domains).
The data for those settings is loaded from the "additional_field" JSON field in the Filter table.

This commit specifically adds an option to match a domain exactly, without matching subdomains.

There are several benefits to using Pydantic:
- Built-in serialization to, and de-serialization from JSON, which is useful for reading and writing to the DB.
- Type validation on initialization.
- By default ignores extra fields when de-serializing, which can prevent things breaking if the extra fields change.