Limit the ext cog to 1 action at a time
Now that cog loading actually has lots of logic in it, rather than spawning tasks to do the actual load work, loading cogs can take time.

Previously we gave no feedback to the caller that something was actually happening. This could mean the caller would try to run the command again, possibly getting the cogs into a bad state.

This change both enforces only 1 action at a time, and gives the caller feedback that their action is in progress.